Refactor: flyte alle transactional funksjoner som har med 'tilstand_virksomhet' og 'tilstand_automatisk_oppdatering' å gjøre, fra Transaction til en ny TilstandVirksomhetTransactional klasse
