+++ src/main/kotlin/no/nav/lydia/tilstandsmaskin/sideeffect/AvsluttVurderingSideEffect.kt
@@ -21,11 +21,11 @@ import no.nav.lydia.tilstandsmaskin.VirksomhetIATilstand
 import no.nav.lydia.tilstandsmaskin.hendelse.GjørVirksomhetKlarTilNyVurdering
 import no.nav.lydia.tilstandsmaskin.hendelse.VurderVirksomhet
 import no.nav.lydia.tilstandsmaskin.hentSisteIASakDto
-import no.nav.lydia.tilstandsmaskin.lagreEllerOppdaterVirksomhetTilstand
 import no.nav.lydia.tilstandsmaskin.lagreHendelse
 import no.nav.lydia.tilstandsmaskin.lagreÅrsakForHendelse
 import no.nav.lydia.tilstandsmaskin.oppdaterStatusPåSak
-import no.nav.lydia.tilstandsmaskin.opprettAutomatiskOppdatering
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.lagreEllerOppdaterVirksomhetTilstand
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.opprettAutomatiskOppdatering
 import java.time.LocalDate
 import java.time.LocalDateTime
