+++ src/main/kotlin/no/nav/lydia/tilstandsmaskin/sideeffect/VurderVirksomhetSideEffect.kt
@@ -17,13 +17,13 @@ import no.nav.lydia.tilgangskontroll.fia.NavAnsatt
 import no.nav.lydia.tilstandsmaskin.NyFlytService
 import no.nav.lydia.tilstandsmaskin.Transaction
 import no.nav.lydia.tilstandsmaskin.VirksomhetIATilstand
-import no.nav.lydia.tilstandsmaskin.lagreEllerOppdaterVirksomhetTilstand
 import no.nav.lydia.tilstandsmaskin.lagreHendelse
 import no.nav.lydia.tilstandsmaskin.lagreÅrsakForHendelse
 import no.nav.lydia.tilstandsmaskin.nyHendelseBasertPåSak
 import no.nav.lydia.tilstandsmaskin.oppdaterStatusPåSak
 import no.nav.lydia.tilstandsmaskin.opprettSak
-import no.nav.lydia.tilstandsmaskin.slettVirksomhetTilstandAutomatiskOppdatering
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.lagreEllerOppdaterVirksomhetTilstand
+import no.nav.lydia.tilstandsmaskin.sideeffect.transactional.TilstandVirksomhetTransactional.Companion.slettVirksomhetTilstandAutomatiskOppdatering
 
 class VurderVirksomhetSideEffect(
     val orgnummer: String,
